+++ codigo/src/concepts/image_modifier/special_operations/access.cpp
@@ -1,8 +1,12 @@
 #include "../../../../headers/concepts/image_modifier.hpp"
 #include <SDL2/SDL_stdinc.h>
+#include <cstdint>
 
 SDL_Color ImageModifier::operator() (uint32_t i, uint32_t j) {
-  if (i < this->texture->h && j < this->texture->w && this->texture->pixels != nullptr) {
+  if (i < static_cast<uint32_t>(this->texture->h) && 
+      j < static_cast<uint32_t>(this->texture->w) && 
+      this->texture->pixels != nullptr
+    ) {
     SDL_Color color;
     SDL_PixelFormat* format = SDL_AllocFormat(SDL_PIXELFORMAT_RGBA8888);
     SDL_GetRGBA(
