+++ codigo/src/primitives/file_processing/zlib_inflate/read_bytes.cpp
@@ -6,7 +6,7 @@ template <bool LSB = true>
 uint64_t read_bytes (uint8_t* vec, uint64_t& pos, uint16_t N) {
   uint64_t ret = 0;
   ret = vec[pos];
-  for (int i = pos + 1; i < N + pos; i++)
+  for (uint16_t i = pos + 1; i < N + pos; i++)
     ret = ret << 8 | vec[i];
   pos += N;
   return ret;
